+++ app/bootstrap/middleware.php
@@ -1,3 +1,8 @@
 <?php
 
 // Global Middlewares go here
+$app->add(new \Slim\Middleware\Session([
+    'name' => $container->settings['framework'],
+    'autorefresh' => true,
+    'lifetime' => '1 hour'
+]));