+++ public/index.php
@@ -1,8 +1,4 @@
 <?php
-session_set_cookie_params(null, null, null, null, true);
-session_start();
-
-// Load Vendor Autoload
 require __DIR__ . '/../vendor/autoload.php';
 
 // Load Settings